Refactored examples into separate components. Added example.
