+++ example/common.jsx
@@ -1,0 +1,7 @@
+export const modifiers = {
+    preventOverflow: {
+        enabled: true,
+        escapeWithReference: true,
+        boundariesElement: 'scrollParent'
+    }
+}
