+++ package.json
@@ -4,14 +4,16 @@
   "description": "React wrapper around PopperJS.",
   "main": "lib/react-popper.js",
   "types": "react-popper.d.ts",
-  "files": ["dist", "lib", "react-popper.d.ts"],
+  "files": [
+    "dist",
+    "lib",
+    "react-popper.d.ts"
+  ],
   "scripts": {
     "build:lib": "babel src --out-dir lib",
-    "build":
-      "npm run build:lib && cross-env NODE_ENV=production webpack --config webpack.prod.config.js",
+    "build": "npm run build:lib && cross-env NODE_ENV=production webpack --config webpack.prod.config.js",
     "dev": "webpack-dev-server --hot --progress --host 0.0.0.0",
-    "postbuild":
-      "cross-env NODE_ENV=production && cross-env TARGET=minify webpack --config webpack.prod.config.js",
+    "postbuild": "cross-env NODE_ENV=production && cross-env TARGET=minify webpack --config webpack.prod.config.js",
     "prebuild": "rimraf dist && mkdir dist",
     "prepublish": "npm run build"
   },
@@ -28,8 +30,7 @@
     "tooltip",
     "popover"
   ],
-  "author":
-    "Travis Arnold <travis@souporserious.com> (http://souporserious.com)",
+  "author": "Travis Arnold <travis@souporserious.com> (http://souporserious.com)",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/souporserious/react-popper/issues"
